Continuing developing contexmenuevent(). Added some debug code
